Add subnavigation for AidApplication to show Assister workflow
