+++ app/controllers/aid_applications/base_controller.rb
@@ -1,0 +1,8 @@
+module AidApplications
+  class BaseController < ApplicationController
+    def current_aid_application
+      current_organization.aid_applications.find(params[:aid_application_id])
+    end
+    helper_method :current_aid_application
+  end
+end
