+++ app/controllers/aid_applications/verifications_controller.rb
@@ -1,0 +1,11 @@
+module AidApplications
+  class VerificationsController < BaseController
+    def edit
+      @aid_application = current_aid_application
+    end
+
+    def update
+      @aid_application = current_aid_application
+    end
+  end
+end
