+++ app/views/aid_applications/disbursements/edit.html.erb
@@ -1,0 +1,3 @@
+<%= render 'aid_applications/navigation' %>
+
+Placeholder for CBO to enter EBT card number that is being given to client.
