+++ app/views/aid_applications/verifications/edit.html.erb
@@ -1,0 +1,3 @@
+<%= render 'aid_applications/navigation' %>
+
+Placeholder to walk person through process of looking at client documents, and also a place to surface potential duplicate household members.
